Rename datetime helper module
